training-studio-1.0.0/VideoAssistant/requirements.txt
jupyter
ffmpeg-python
python-dotenv
shiny
git+https://github.com/jcheng5/shinymedia
pandas
matplotlib
groq
gTTS
pyttsx3